Fix to configure path to bin/lib of postgresql92 and pgpoolAdmin tar ball

+++ redhat/rpm_installer/getsources.sh
@@ -1,21 +1,38 @@
 #!/bin/sh
 
+# ---------------------------------------------------------------------
+# configuration
+# ---------------------------------------------------------------------
+
+git_dir=$HOME/git
+
+## pgpool-II
+pgpool_src_dir=$git_dir/pgpool2
+pgpool_tarball_dir=$pgpool_src_dir
+pgpool_version=3.3.0
+## pgpoolAdmin
+admin_src_dir=$git_dir/pgpooladmin
+admin_tarball_dir=$admin_src_dir/tools
+admin_version=3.3
+
 username=`whoami`
 rpm_dir=$HOME/rpm
-git_dir=$HOME/git
-pgpool_src_dir=$git_dir/pgpool2/master
-admin_src_dir=$git_dir/pgpooladmin
-pgdata_template=$HOME/local/pgsql/data
 work_dir=`pwd`/work
+
+## postgresql92
+bin_path=/usr/pgsql-9.2/bin
+lib_path=/usr/pgsql-9.2/bin
+export PATH=$PATH:$bin_path
+export LD_LIBRARY=$LD_LIBRARY:$lib_path
 
 echo "* Setup starts."
 echo
 
 # ---------------------------------------------------------------------
-# setup for rpmnbuild
+# setup for rpmbuild
 # ---------------------------------------------------------------------
 
-echo -n "    Setup for rpmbild ... "
+echo -n "    Setup for rpmbuild ... "
 
 mkdir -p $rpm_dir/{BUILD,RPMS,SOURCES,SPECS,SRPMS}
 
@@ -47,10 +64,21 @@ cp -f $admin_src_dir/pgpoolAdmin.spec $work_dir
 cd $pgpool_src_dir
 make dist > /dev/null 2>&1
 if [ $? -ne 0 ]; then
-    "Failed."
+    echo "Failed."
     exit
 fi
-mv pgpool-II-*.tar.gz $rpm_dir/SOURCES
+if [ ! -f $pgpool_tarball_dir/pgpool-II-$pgpool_version.tar.gz ]; then
+    echo "$pgpool_tarball_dir/pgpool-II-$pgpool_version.tar.gz not found."
+	exit
+fi
+mv pgpool-II-$pgpool_version.tar.gz $rpm_dir/SOURCES
+
+# pgpoolAdmin-*.tar.gz
+if [ ! -f $admin_tarball_dir/pgpoolAdmin-$admin_version.tar.gz ]; then
+    echo "$admin_tarboll_dir/pgpoolAdmin-$admin_version.tar.gz not found."
+	exit
+fi
+cp $admin_src_dir/tools/pgpoolAdmin-$admin_version.tar.gz $rpm_dir/SOURCES
 
 # pgpool.conf.sample.patch, pgpool.init, pgpool.sysconfig
 cp -f $pgpool_src_dir/redhat/pgpool.conf.sample.patch $rpm_dir/SOURCES/
